services/anomaly-detection-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
kafka-python>=2.0.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
